[Consumer] move out a few more internals
